Moved PeerId and PeerEndPoint related classes to network package, created MessageDispatcher interface and renamed the implementation to HashMapMessageDispatcher
