+++ app/src/main/java/org/albard/dubito/app/messaging/messages/GameMessageBase.java
@@ -2,7 +2,7 @@ package org.albard.dubito.app.messaging.messages;
 
 import java.util.Set;
 
-import org.albard.dubito.app.PeerId;
+import org.albard.dubito.app.network.PeerId;
 
 public abstract class GameMessageBase implements GameMessage {
     private final PeerId sender;
