+++ app/src/main/java/org/albard/dubito/app/network/PeerEndPoint.java
@@ -1,4 +1,4 @@
-package org.albard.dubito.app;
+package org.albard.dubito.app.network;
 
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
